Utilize ActionCable backend
Use the websockets to push changes to clients. This is a very simple implementation; it only reports the change and then forwards it to the subscribed components. In this case the pizza controller. Which then fetches the pizzas from the http backend. It would be an improvement to transmit the change through the cable and apply that to the model.